+++ avl-core/src/test/resources/audit-fixtures/TestNgSampleTest.java
@@ -1,0 +1,11 @@
+package fixtures;
+
+import io.github.byreshb.avl.annotation.Team;
+import org.testng.annotations.Test;
+
+public class TestNgSampleTest {
+
+  @Team("platform")
+  @Test
+  public void missingLayerOnly() {}
+}